requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
itsdangerous>=2.1
aiofiles>=23.0
deepface>=0.0.89
opencv-python-headless>=4.9
tf-keras>=2.16
retina-face>=0.0.17
numpy>=1.26
fpdf2>=2.7
